Move sdk.data_pipeline to sdk.actor.data_sourcing
All the types left in `frequenz.sdk.data_pipeline(.types)` were specific
to the data sourcing actor, so they were moved to
`frequenz.sdk.data_sourcing.types` (and exported in
`frequenz.sdk.data_sourcing`).

The resampling actor for now is also forwarding requests to the data
sourcing actor, so it also re-exports those symbols.
