+++ examples/sdk_resampling_example.py
@@ -13,8 +13,11 @@ from frequenz.channels import Broadcast, MergeNamed
 
 from frequenz.sdk.actor import ChannelRegistry
 from frequenz.sdk.actor.data_sourcing import DataSourcingActor
-from frequenz.sdk.actor.resampling import ComponentMetricsResamplingActor
-from frequenz.sdk.data_pipeline import ComponentMetricId, ComponentMetricRequest
+from frequenz.sdk.actor.resampling import (
+    ComponentMetricId,
+    ComponentMetricRequest,
+    ComponentMetricsResamplingActor,
+)
 from frequenz.sdk.microgrid import ComponentCategory, microgrid_api
 
 HOST = "microgrid.sandbox.api.frequenz.io"
