--- src/frequenz/sdk/data_pipeline/__init__.py
@@ -1,15 +1,0 @@
-"""Utils for the Data Pipeline.
-
-Copyright
-Copyright © 2022 Frequenz Energy-as-a-Service GmbH
-
-License
-MIT
-"""
-
-from .types import ComponentMetricId, ComponentMetricRequest
-
-__all__ = [
-    "ComponentMetricRequest",
-    "ComponentMetricId",
-]
